WIP: Delete File Confirmation

+++ src/client/Sidebar/FileListing.tsx
@@ -24,7 +24,10 @@ export const FileListing = ({
 
   const handleDeleteClick = useCallback(
     (event) => {
-      handleDeleteFileClick(fileId, event);
+      const x = window.confirm("Are you sure you want to delete?");
+      if( x === true ){
+        handleDeleteFileClick(fileId, event);
+      }
     },
     [fileId, handleDeleteFileClick],
   );
